add command class, refactor command actions into command
